refactor protocol name
